refactor: merge download/icon tests into existing service test files
